# openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GPT Backend API",
    "description": "Backend API for Gmail management and app control",
    "version": "1.0.0"
  },
  "paths": {
    "/": {
      "get": {
        "summary": "Root",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/gmail/messages": {
      "get": {
        "summary": "Get Messages",
        "description": "Get Gmail messages",
        "operationId": "get_messages_gmail_messages_get",
        "parameters": [
          {
            "name": "max_results",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Max Results"
            }
          },
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageResponse"
                  },
                  "title": "Response Get Messages Gmail Messages Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/gmail/messages/{message_id}": {
      "get": {
        "summary": "Get Message",
        "description": "Get a specific Gmail message by ID",
        "operationId": "get_message_gmail_messages__message_id__get",
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Message Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/gmail/send": {
      "post": {
        "summary": "Send Email",
        "description": "Send a new email",
        "operationId": "send_email_gmail_send_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendEmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/gmail/reply": {
      "post": {
        "summary": "Reply Email",
        "description": "Reply to an email",
        "operationId": "reply_email_gmail_reply_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplyEmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/gmail/auth/status": {
      "get": {
        "summary": "Auth Status",
        "description": "Check Gmail authentication status",
        "operationId": "auth_status_gmail_auth_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/gmail/auth/url": {
      "get": {
        "summary": "Get Auth Url",
        "description": "Get Gmail OAuth authorization URL",
        "operationId": "get_auth_url_gmail_auth_url_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/gmail/auth/callback": {
      "post": {
        "summary": "Auth Callback",
        "description": "Handle OAuth callback and store credentials",
        "operationId": "auth_callback_gmail_auth_callback_post",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "redirect_uri",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Redirect Uri"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/oauth2callback": {
      "get": {
        "summary": "Oauth2 Callback",
        "description": "OAuth2 callback endpoint for web applications",
        "operationId": "oauth2_callback_oauth2callback_get",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "name": "error",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/apps/control": {
      "post": {
        "summary": "Control App",
        "description": "Start or stop an application",
        "operationId": "control_app_apps_control_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppControlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppControlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/apps/list": {
      "get": {
        "summary": "List Apps",
        "description": "List available apps that can be controlled",
        "operationId": "list_apps_apps_list_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AppControlRequest": {
        "properties": {
          "app_name": {
            "type": "string",
            "title": "App Name"
          },
          "action": {
            "type": "string",
            "title": "Action"
          }
        },
        "type": "object",
        "required": [
          "app_name",
          "action"
        ],
        "title": "AppControlRequest"
      },
      "AppControlResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "app_name": {
            "type": "string",
            "title": "App Name"
          },
          "action": {
            "type": "string",
            "title": "Action"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message",
          "app_name",
          "action"
        ],
        "title": "AppControlResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MessageResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "from_email": {
            "type": "string",
            "title": "From Email"
          },
          "to": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "To"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "date": {
            "type": "string",
            "title": "Date"
          },
          "snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snippet"
          }
        },
        "type": "object",
        "required": [
          "id",
          "thread_id",
          "from_email",
          "to",
          "subject",
          "body",
          "date"
        ],
        "title": "MessageResponse"
      },
      "ReplyEmailRequest": {
        "properties": {
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "in_reply_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "In Reply To"
          }
        },
        "type": "object",
        "required": [
          "thread_id",
          "body"
        ],
        "title": "ReplyEmailRequest"
      },
      "SendEmailRequest": {
        "properties": {
          "to": {
            "type": "string",
            "format": "email",
            "title": "To"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "cc": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "email"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cc"
          },
          "bcc": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "email"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bcc"
          }
        },
        "type": "object",
        "required": [
          "to",
          "subject",
          "body"
        ],
        "title": "SendEmailRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}